Module catalog::helper

Macros

define_catalog_value 🔒

Structs

CATALOG_KEY_PATTERN 🔒

CatalogKey

CatalogValue

SCHEMA_KEY_PATTERN 🔒

SchemaKey

SchemaValue

TABLE_GLOBAL_KEY_PATTERN 🔒

TABLE_REGIONAL_KEY_PATTERN 🔒

TableGlobalKey

Table global info has only one key across all datanodes so it does not have node_id field.

TableGlobalValue

Table global info contains necessary info for a datanode to create table regions, including table id, table meta(schema…), region id allocation across datanodes.

TableRegionalKey

Table regional info that varies between datanode, so it contains a node_id field.

TableRegionalValue

Regional table info of specific datanode, including table version on that datanode and region ids allocated by metasrv.

__FieldVisitor 🔒

__FieldVisitor 🔒

__Visitor 🔒

__Visitor 🔒

__Visitor 🔒

__Visitor 🔒

Enums

__Field 🔒

__Field 🔒

Constants

ALPHANUMERICS_NAME_PATTERN 🔒

CATALOG_KEY_PREFIX

FIELDS 🔒

FIELDS 🔒

SCHEMA_KEY_PREFIX

TABLE_GLOBAL_KEY_PREFIX

TABLE_REGIONAL_KEY_PREFIX

Statics

LAZY 🔒

LAZY 🔒

LAZY 🔒

LAZY 🔒

Functions

__stability 🔒

__stability 🔒

__stability 🔒

__stability 🔒

__static_ref_initialize 🔒

__static_ref_initialize 🔒

__static_ref_initialize 🔒

__static_ref_initialize 🔒

build_catalog_prefix

build_schema_prefix

build_table_global_prefix

build_table_regional_prefix