Struct catalog::local::manager::LocalCatalogManager

pub struct LocalCatalogManager {
    system: Arc<SystemCatalog>,
    catalogs: Arc<MemoryCatalogManager>,
    engine_manager: TableEngineManagerRef,
    next_table_id: AtomicU32,
    init_lock: Mutex<bool>,
    register_lock: Mutex<()>,
    system_table_requests: Mutex<Vec<RegisterSystemTableRequest>>,
}

A CatalogManager consists of a system catalog and a bunch of user catalogs.

Fields
system: Arc<SystemCatalog>catalogs: Arc<MemoryCatalogManager>engine_manager: TableEngineManagerRefnext_table_id: AtomicU32init_lock: Mutex<bool>register_lock: Mutex<()>system_table_requests: Mutex<Vec<RegisterSystemTableRequest>>
Implementations

impl LocalCatalogManager

pub async fn try_new(engine_manager: TableEngineManagerRef) -> Result<Self>

Create a new CatalogManager with given user catalogs and mito engine

pub async fn init(&self) -> Result<()>

Scan all entries from system catalog table

async fn init_system_catalog(&self) -> Result<()>

async fn collect_system_catalog_entries( &self, stream: SendableRecordBatchStream ) -> Result<Vec<Entry>>

Collect stream of system catalog entries to Vec<Entry>

fn record_batch_to_entry(rb: RecordBatch) -> Result<Vec<Entry>>

Convert RecordBatch to a vector of Entry.

async fn handle_system_catalog_entries( &self, entries: Vec<Entry> ) -> Result<TableId>

Processes records from system catalog table and returns the max table id persisted in system catalog table.

fn sort_entries(entries: Vec<Entry>) -> Vec<Entry>

Sort catalog entries to ensure catalog entries comes first, then schema entries, and table entries is the last.

async fn open_and_register_table(&self, t: &TableEntry) -> Result<()>

async fn check_state(&self) -> Result<()>

async fn check_catalog_schema_exist( &self, catalog_name: &str, schema_name: &str ) -> Result<()>