catalog::system_schema::information_schema

Trait InformationSchemaTableFactory

pub trait InformationSchemaTableFactory {
    // Required method
    fn make_information_table(
        &self,
        req: MakeInformationTableRequest,
    ) -> SystemTableRef;
}

A factory trait for making information schema tables.

This trait allows for extensibility of the information schema by providing a way to dynamically create custom information schema tables.

Required Methods

fn make_information_table( &self, req: MakeInformationTableRequest, ) -> SystemTableRef