Trait catalog::CatalogManager

pub trait CatalogManager: Send + Sync {

    // Required methods
    fn as_any(&self) -> &dyn Any;
    fn start<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn register_catalog<'life0, 'async_trait>(
        &'life0 self,
        name: String
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn register_schema<'life0, 'async_trait>(
        &'life0 self,
        request: RegisterSchemaRequest
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn deregister_schema<'life0, 'async_trait>(
        &'life0 self,
        request: DeregisterSchemaRequest
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn register_table<'life0, 'async_trait>(
        &'life0 self,
        request: RegisterTableRequest
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn deregister_table<'life0, 'async_trait>(
        &'life0 self,
        request: DeregisterTableRequest
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn rename_table<'life0, 'async_trait>(
        &'life0 self,
        request: RenameTableRequest
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn register_system_table<'life0, 'async_trait>(
        &'life0 self,
        request: RegisterSystemTableRequest
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn catalog_names<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn schema_names<'life0, 'life1, 'async_trait>(
        &'life0 self,
        catalog: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Vec<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn table_names<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        catalog: &'life1 str,
        schema: &'life2 str
    ) -> Pin<Box<dyn Future<Output = Result<Vec<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn catalog_exist<'life0, 'life1, 'async_trait>(
        &'life0 self,
        catalog: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn schema_exist<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        catalog: &'life1 str,
        schema: &'life2 str
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn table_exist<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        catalog: &'life1 str,
        schema: &'life2 str,
        table: &'life3 str
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
    fn table<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        catalog: &'life1 str,
        schema: &'life2 str,
        table_name: &'life3 str
    ) -> Pin<Box<dyn Future<Output = Result<Option<TableRef>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;

}

Required Methods

fn as_any(&self) -> &dyn Any

fn start<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Starts a catalog manager.

fn register_catalog<'life0, 'async_trait>( &'life0 self, name: String ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Registers a catalog to catalog manager, returns whether the catalog exist before.

fn register_schema<'life0, 'async_trait>( &'life0 self, request: RegisterSchemaRequest ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Register a schema with catalog name and schema name. Retuens whether the schema registered.

Errors

This method will/should fail if catalog not exist

fn deregister_schema<'life0, 'async_trait>( &'life0 self, request: DeregisterSchemaRequest ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Deregisters a database within given catalog/schema to catalog manager

fn register_table<'life0, 'async_trait>( &'life0 self, request: RegisterTableRequest ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Registers a table within given catalog/schema to catalog manager, returns whether the table registered.

Errors

This method will/should fail if catalog or schema not exist

fn deregister_table<'life0, 'async_trait>( &'life0 self, request: DeregisterTableRequest ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Deregisters a table within given catalog/schema to catalog manager

fn rename_table<'life0, 'async_trait>( &'life0 self, request: RenameTableRequest ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Rename a table to RenameTableRequest::new_table_name, returns whether the table is renamed.

fn register_system_table<'life0, 'async_trait>( &'life0 self, request: RegisterSystemTableRequest ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Register a system table, should be called before starting the manager.

fn catalog_names<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<String>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn schema_names<'life0, 'life1, 'async_trait>( &'life0 self, catalog: &'life1 str ) -> Pin<Box<dyn Future<Output = Result<Vec<String>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn table_names<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, catalog: &'life1 str, schema: &'life2 str ) -> Pin<Box<dyn Future<Output = Result<Vec<String>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

fn catalog_exist<'life0, 'life1, 'async_trait>( &'life0 self, catalog: &'life1 str ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn schema_exist<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, catalog: &'life1 str, schema: &'life2 str ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

fn table_exist<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, catalog: &'life1 str, schema: &'life2 str, table: &'life3 str ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

fn table<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, catalog: &'life1 str, schema: &'life2 str, table_name: &'life3 str ) -> Pin<Box<dyn Future<Output = Result<Option<TableRef>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Returns the table by catalog, schema and table name.