Module catalog::system_schema::pg_catalog

Modules

pg_catalog_memory_table 🔒

pg_class 🔒

pg_namespace 🔒

table_names 🔒

Macros

setup_memory_table 🔒

Structs

MEMORY_TABLES 🔒

PGCatalogProvider

PGCatalogProvider is the provider for a schema named pg_catalog, it is not a catalog.

Constants

OID_COLUMN_NAME 🔒

The column name for the OID column. The OID column is a unique identifier of type u32 for each object in the database.

PG_CLASS

PG_DATABASE

PG_NAMESPACE

PG_TYPE

Functions

oid_column 🔒