Trait catalog::CatalogManager

pub trait CatalogManager: CatalogList {
    // Required methods
    fn start<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn register_table<'life0, 'async_trait>(
        &'life0 self,
        request: RegisterTableRequest
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn deregister_table<'life0, 'async_trait>(
        &'life0 self,
        request: DeregisterTableRequest
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn register_schema<'life0, 'async_trait>(
        &'life0 self,
        request: RegisterSchemaRequest
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn rename_table<'life0, 'async_trait>(
        &'life0 self,
        request: RenameTableRequest
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn register_system_table<'life0, 'async_trait>(
        &'life0 self,
        request: RegisterSystemTableRequest
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn schema(
        &self,
        catalog: &str,
        schema: &str
    ) -> Result<Option<SchemaProviderRef>>;
    fn table<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        catalog: &'life1 str,
        schema: &'life2 str,
        table_name: &'life3 str
    ) -> Pin<Box<dyn Future<Output = Result<Option<TableRef>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
}

Required Methods

fn start<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Starts a catalog manager.

fn register_table<'life0, 'async_trait>( &'life0 self, request: RegisterTableRequest ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Registers a table within given catalog/schema to catalog manager, returns whether the table registered.

fn deregister_table<'life0, 'async_trait>( &'life0 self, request: DeregisterTableRequest ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Deregisters a table within given catalog/schema to catalog manager, returns whether the table deregistered.

fn register_schema<'life0, 'async_trait>( &'life0 self, request: RegisterSchemaRequest ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Register a schema with catalog name and schema name. Retuens whether the schema registered.

fn rename_table<'life0, 'async_trait>( &'life0 self, request: RenameTableRequest ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Rename a table to RenameTableRequest::new_table_name, returns whether the table is renamed.

fn register_system_table<'life0, 'async_trait>( &'life0 self, request: RegisterSystemTableRequest ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Register a system table, should be called before starting the manager.

fn schema(&self, catalog: &str, schema: &str) -> Result<Option<SchemaProviderRef>>

fn table<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, catalog: &'life1 str, schema: &'life2 str, table_name: &'life3 str ) -> Pin<Box<dyn Future<Output = Result<Option<TableRef>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Returns the table by catalog, schema and table name.